refactor: separate select tests into a new file
